refactor serde for conditional expressions
